requirements.txt
streamlit
pandas
matplotlib
numpy
praat-parselmouth
textgrid